refactor: classify auth api
