+++ apps/api/src/auth/guard/anonymous.strategy.ts
