+++ apps/api/src/auth/guard/jwt.strategy.ts
@@ -3,12 +3,12 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { JwtPayload } from 'jsonwebtoken';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { UserService } from '../user.service';
+import { AuthService } from '../auth.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
   constructor(
-    private readonly userService: UserService,
+    private readonly authService: AuthService,
     private readonly customConfigService: CustomConfigService,
   ) {
     super({
@@ -21,7 +21,7 @@ export class JwtStrategy extends PassportStrategy(Strategy, 'jwt') {
 
   async validate({ sub }: JwtPayload) {
     if (!sub) throw new UnauthorizedException('invalid token');
-    return this.userService.findUserByUuid(sub).catch(() => {
+    return await this.authService.findUserByUuid(sub).catch(() => {
       throw new UnauthorizedException();
     });
   }
