+++ apps/api/src/auth/guard/jwtOptional.strategy.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
-import { UserService } from '../user.service';
+import { AuthService } from '../auth.service';
 import { JwtPayload } from 'jsonwebtoken';
 import { User } from '@prisma/client';
 import { CustomConfigService } from '@lib/custom-config';
@@ -12,7 +12,7 @@ export class JwtOptionalStrategy extends PassportStrategy(
   'jwt-optional',
 ) {
   constructor(
-    private readonly userService: UserService,
+    private readonly authService: AuthService,
     private readonly customConfigService: CustomConfigService,
   ) {
     super({
@@ -25,7 +25,7 @@ export class JwtOptionalStrategy extends PassportStrategy(
 
   async validate({ sub }: JwtPayload): Promise<User | void> {
     if (!sub) return undefined;
-    return this.userService.findUserByUuid(sub).catch(() => {
+    return this.authService.findUserByUuid(sub).catch(() => {
       return undefined;
     });
   }
