+++ apps/api/src/auth/types/jwtToken.type.ts
@@ -1,0 +1,10 @@
+export type JwtTokenType = {
+  access_token: string;
+  refresh_token: string;
+  consent_required: boolean;
+};
+
+export type IssueTokenType = {
+  access_token: string;
+  refresh_token: string;
+};
