+++ apps/api/src/notice/notice.controller.ts
@@ -40,7 +40,7 @@ import { GetUser } from '../user/decorator/get-user.decorator';
 import { GroupsGuard } from '@lib/infoteam-groups/guard/groups.guard';
 import { GetGroups } from '../user/decorator/get-groups.decorator';
 import { GroupsUserInfo } from '@lib/infoteam-groups/types/groups.type';
-import { JwtGuard } from '../user/guard/jwt.guard';
+import { JwtGuard } from '../auth/guard/jwt.guard';
 
 @ApiTags('notice')
 @ApiBearerAuth('jwt')
