--- apps/api/src/user/dto/req/logout.dto.ts
@@ -1,6 +1,0 @@
-import { IsString } from 'class-validator';
-
-export class LogoutDto {
-  @IsString()
-  access_token: string;
-}
