+++ apps/api/src/user/user.controller.ts
@@ -6,21 +6,15 @@ import {
   HttpCode,
   HttpStatus,
   Post,
-  Req,
-  Res,
-  UnauthorizedException,
   UseGuards,
   UsePipes,
   ValidationPipe,
 } from '@nestjs/common';
-import { Request, Response } from 'express';
-import { JwtToken } from './dto/res/jwtToken.dto';
 import { UserService } from './user.service';
 import {
   ApiBearerAuth,
   ApiCreatedResponse,
   ApiInternalServerErrorResponse,
-  ApiOAuth2,
   ApiOkResponse,
   ApiOperation,
   ApiTags,
@@ -31,84 +25,13 @@ import { GetUser } from './decorator/get-user.decorator';
 import { UserInfoRes } from './dto/res/userInfoRes.dto';
 import { setFcmTokenRes } from './dto/res/setFcmTokenRes.dto';
 import { setFcmTokenReq } from './dto/req/setFcmTokenReq.dto';
-import { JwtGuard, JwtOptionalGuard } from './guard/jwt.guard';
-import ms, { StringValue } from 'ms';
-import { CustomConfigService } from '@lib/custom-config';
+import { JwtGuard, JwtOptionalGuard } from '../auth/guard/jwt.guard';
 
 @ApiTags('user')
 @Controller('user')
 @UsePipes(ValidationPipe)
 export class UserController {
-  private readonly refreshTokenExpire: number;
-  constructor(
-    private readonly userService: UserService,
-    private readonly customConfigService: CustomConfigService,
-  ) {
-    this.refreshTokenExpire = ms(
-      customConfigService.REFRESH_TOKEN_EXPIRE as StringValue,
-    );
-  }
-
-  @ApiOperation({
-    summary: 'Login',
-    description: 'Issue ziggle JWT token',
-  })
-  @ApiOkResponse({ description: 'Return jwt token' })
-  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
-  @ApiInternalServerErrorResponse({ description: 'Internal Server Error' })
-  @ApiOAuth2(['email', 'profile', 'openid'], 'oauth2')
-  @Post('login')
-  async login(
-    @Req() req: Request,
-    @Res({ passthrough: true }) res: Response,
-  ): Promise<JwtToken> {
-    const auth = req.headers['authorization'];
-    if (!auth) throw new UnauthorizedException();
-    const { access_token, refresh_token, consent_required } =
-      await this.userService.login(auth);
-    res.cookie('refresh_token', refresh_token, {
-      httpOnly: true,
-      secure: true,
-      sameSite: 'strict',
-      expires: new Date(Date.now() + this.refreshTokenExpire),
-      path: '/user',
-    });
-    return { access_token, consent_required };
-  }
-
-  @ApiOperation({
-    summary: 'Refresh token',
-    description: 'Refresh the access token from idp',
-  })
-  @ApiCreatedResponse({ type: JwtToken, description: 'Return jwt token' })
-  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
-  @ApiInternalServerErrorResponse({ description: 'Internal Server Error' })
-  @Post('refresh')
-  async refreshToken(@Req() req: Request): Promise<JwtToken> {
-    const refreshToken = req.cookies['refresh_token'];
-    if (!refreshToken) throw new UnauthorizedException();
-    const tokens = await this.userService.refresh(refreshToken);
-    return tokens;
-  }
-
-  @ApiOperation({
-    summary: 'Logout',
-    description: 'Logout the user from the cookie and idp',
-  })
-  @ApiCreatedResponse({ description: 'Return jwt token' })
-  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
-  @ApiInternalServerErrorResponse({ description: 'Internal Server Error' })
-  @ApiBearerAuth('jwt')
-  @Post('logout')
-  @UseGuards(JwtGuard)
-  async logout(
-    @Req() req: Request,
-    @Res({ passthrough: true }) res: Response,
-  ): Promise<void> {
-    const refreshToken = req.cookies['refresh_token'];
-    res.clearCookie('refresh_token');
-    return this.userService.logout(refreshToken);
-  }
+  constructor(private readonly userService: UserService) {}
 
   @ApiOperation({
     summary: 'post consent',
